Refactor code to stop using script for starting
