+++ Makefile
@@ -4,34 +4,27 @@ flags=-ldflags="-s -w -X main.version=${VERSION}"
 all: build
 
 build:
-	go clean; rm -rf pkg; go build -o udp_server ${flags} udp_server.go
-	go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg; go build -o udp_collector ${flags}
 
 build_debug:
-	go clean; rm -rf pkg; go build -o udp_server ${flags} -gcflags="-m -m" udp_server.go
-	go build -o udp_server_monitor ${flags} -gcflags="-m -m" udp_server_monitor.go
+	go clean; rm -rf pkg; go build -o udp_collector ${flags} -gcflags="-m -m"
 
 build_all: build_osx build_linux build
 
 build_osx:
-	go clean; rm -rf pkg udp_server_osx; GOOS=darwin go build -o udp_server ${flags} udp_server.go
-	GOOS=darwin go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg udp_collector_osx; GOOS=darwin go build -o udp_collector ${flags}
 
 build_linux:
-	go clean; rm -rf pkg udp_server_linux; GOOS=linux go build -o udp_server ${flags} udp_server.go
-	GOOS=linux go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg udp_collector_linux; GOOS=linux go build -o udp_collector ${flags}
 
 build_power8:
-	go clean; rm -rf pkg udp_server_power8; GOARCH=ppc64le GOOS=linux go build -o udp_server ${flags} udp_server.go
-	GOARCH=ppc64le GOOS=linux go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg udp_collector_power8; GOARCH=ppc64le GOOS=linux go build -o udp_collector ${flags}
 
 build_arm64:
-	go clean; rm -rf pkg udp_server_arm64; GOARCH=arm64 GOOS=linux go build -o udp_server ${flags} udp_server.go
-	GOARCH=arm64 GOOS=linux go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg udp_collector_arm64; GOARCH=arm64 GOOS=linux go build -o udp_collector ${flags}
 
 build_windows:
-	go clean; rm -rf pkg udp_server.exe; GOARCH=amd64 GOOS=windows go build -o udp_server ${flags} udp_server.go
-	GOARCH=amd64 GOOS=windows go build -o udp_server_monitor ${flags} udp_server_monitor.go
+	go clean; rm -rf pkg udp_collector.exe; GOARCH=amd64 GOOS=windows go build -o udp_collector ${flags}
 
 install:
 	go install
